renamed files to match the classes names inside
