+++ src/emails/application/exercise_email_builder.py
